+++ src/lib/components/Button/Button.template.tsx
@@ -11,6 +11,7 @@ export const ButtonTemplate = (props: IButtonProps) => {
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
 		cssClasses.push(`btn`);
+		cssClasses.push(styles.button);
 		if (variant !== 'outline' && variant !== 'text') {
 			cssClasses.push(`btn-${color}`);
 		}