+++ src/lib/components/Modal/ModalFooter.tsx
@@ -1,9 +1,10 @@
 import React, { ReactNode } from 'react';
+import styles from './ModalFooter.module.scss';
 
 interface IModalFooterProps {
 	children?: ReactNode;
 }
 
 export const ModalFooter = ({ children }: IModalFooterProps) => (
-	<div className="modal-body">{children}</div>
+	<div className={"modal-footer " + styles.modalFooter}>{children}</div>
 );