+++ changelog.md
@@ -1,3 +1,10 @@
+## 15.0.2
+- checkbox - label disabled fix
+- dropdown - added scss modules
+- modal - footer style fix
+- deleted global DropDown.scss
+- deleted global Modal.scss
+
 ## 15.0.1
 - style fixes